@@ -4,6 +4,7 @@ namespace App\Console;
 
 use Illuminate\Console\Scheduling\Schedule;
 use Illuminate\Foundation\Console\Kernel as ConsoleKernel;
+use Laravel\Sanctum\Console\Commands\PruneExpired;
 
 class Kernel extends ConsoleKernel
 {
@@ -14,9 +15,9 @@ class Kernel extends ConsoleKernel
      *
      * @return void
      */
-    protected function schedule(Schedule $schedule)
+    protected function schedule(Schedule $schedule): void
     {
-        $schedule->command('sanctum:prune-expired --hours=24')->daily();
+        $schedule->command(PruneExpired::class, ['hours' => 24])->daily();
     }
 
     /**
@@ -24,7 +25,7 @@ class Kernel extends ConsoleKernel
      *
      * @return void
      */
-    protected function commands()
+    protected function commands(): void
     {
         $this->load(__DIR__ . '/Commands');
 
