@@ -5,6 +5,8 @@ namespace App\Mail\User;
 use App\Models\User;
 use Illuminate\Bus\Queueable;
 use Illuminate\Mail\Mailable;
+use Illuminate\Mail\Mailables\Content;
+use Illuminate\Mail\Mailables\Envelope;
 use Illuminate\Queue\SerializesModels;
 
 class PasswordReset extends Mailable
@@ -13,44 +15,54 @@ class PasswordReset extends Mailable
     use SerializesModels;
 
     /**
-     * The user instance.
-     *
-     * @var User
-     */
-    private $user;
-
-    /**
-     * Password reset token.
-     *
-     * @var string
-     */
-    private $token;
-
-    /**
      * Create a new message instance.
      *
-     * @param User $user
+     * @param User   $user
+     * @param string $token
      *
      * @return void
      */
-    public function __construct(User $user, string $token)
-    {
-        $this->user = $user;
-        $this->token = $token;
+    public function __construct(
+        private User $user,
+        private string $token
+    ) {
     }
 
     /**
-     * Build the message.
-     *
-     * @return $this
+     * Get the message envelope.
      */
-    public function build()
+    public function envelope(): Envelope
     {
-        return $this
-            ->subject(strval(__('email.USER.PASSWORD-RESET.SUBJECT')))
-            ->markdown('emails.user.password-reset', [
+        $subject = __('email.USER.PASSWORD-RESET.SUBJECT');
+
+        assert(is_string($subject));
+
+        return new Envelope(
+            subject: $subject,
+        );
+    }
+
+    /**
+     * Get the message content definition.
+     */
+    public function content(): Content
+    {
+        return new Content(
+            markdown: 'emails.user.password-reset',
+            with: [
                 'user'  => $this->user,
                 'token' => $this->token,
-            ]);
+            ]
+        );
+    }
+
+    /**
+     * Get the attachments for the message.
+     *
+     * @return array<int, \Illuminate\Mail\Mailables\Attachment>
+     */
+    public function attachments(): array
+    {
+        return [];
     }
 }
