@@ -14,9 +14,9 @@ class UserPolicy
      *
      * @param \App\Models\User $user
      *
-     * @return bool|\Illuminate\Auth\Access\Response
+     * @return bool
      */
-    public function viewAny(User $user)
+    public function viewAny(User $user): bool
     {
         return true;
     }
@@ -27,9 +27,9 @@ class UserPolicy
      * @param \App\Models\User $user
      * @param \App\Models\User $model
      *
-     * @return bool|\Illuminate\Auth\Access\Response
+     * @return bool
      */
-    public function view(User $user, User $model)
+    public function view(User $user, User $model): bool
     {
         return true;
     }
@@ -39,9 +39,9 @@ class UserPolicy
      *
      * @param \App\Models\User $user
      *
-     * @return bool|\Illuminate\Auth\Access\Response
+     * @return bool
      */
-    public function create(User $user)
+    public function create(User $user): bool
     {
         return false;
     }
@@ -52,9 +52,9 @@ class UserPolicy
      * @param \App\Models\User $user
      * @param \App\Models\User $model
      *
-     * @return bool|\Illuminate\Auth\Access\Response
+     * @return bool
      */
-    public function update(User $user, User $model)
+    public function update(User $user, User $model): bool
     {
         return $user->id === $model->id;
     }
@@ -65,9 +65,9 @@ class UserPolicy
      * @param \App\Models\User $user
      * @param \App\Models\User $model
      *
-     * @return bool|\Illuminate\Auth\Access\Response
+     * @return bool
      */
-    public function delete(User $user, User $model)
+    public function delete(User $user, User $model): bool
     {
         return false;
     }
@@ -78,9 +78,9 @@ class UserPolicy
      * @param \App\Models\User $user
      * @param \App\Models\User $model
      *
-     * @return bool|\Illuminate\Auth\Access\Response
+     * @return bool
      */
-    public function restore(User $user, User $model)
+    public function restore(User $user, User $model): bool
     {
         return false;
     }
@@ -91,9 +91,9 @@ class UserPolicy
      * @param \App\Models\User $user
      * @param \App\Models\User $model
      *
-     * @return bool|\Illuminate\Auth\Access\Response
+     * @return bool
      */
-    public function forceDelete(User $user, User $model)
+    public function forceDelete(User $user, User $model): bool
     {
         return false;
     }
