@@ -12,7 +12,7 @@ return [
     |
     */
 
-    'defaults'         => [
+    'defaults' => [
         'guard'     => 'web',
         'passwords' => 'users',
     ],
@@ -34,7 +34,7 @@ return [
     |
     */
 
-    'guards'           => [
+    'guards' => [
         'web' => [
             'driver'   => 'session',
             'provider' => 'users',
@@ -58,7 +58,7 @@ return [
     |
     */
 
-    'providers'        => [
+    'providers' => [
         'users' => [
             'driver' => 'eloquent',
             'model'  => App\Models\User::class,
@@ -79,16 +79,20 @@ return [
     | than one user table or model in the application and you want to have
     | separate password reset settings based on the specific user types.
     |
-    | The expire time is the number of minutes that each reset token will be
+    | The expiry time is the number of minutes that each reset token will be
     | considered valid. This security feature keeps tokens short-lived so
     | they have less time to be guessed. You may change this as needed.
     |
+    | The throttle setting is the number of seconds a user must wait before
+    | generating more password reset tokens. This prevents the user from
+    | quickly generating a very large amount of password reset tokens.
+    |
     */
 
-    'passwords'        => [
+    'passwords' => [
         'users' => [
             'provider' => 'users',
-            'table'    => 'password_resets',
+            'table'    => 'password_reset_tokens',
             'expire'   => 60,
             'throttle' => 60,
         ],
