@@ -30,24 +30,24 @@ return [
     |
     */
 
-    'stores'  => [
-        'apc'       => [
+    'stores' => [
+        'apc' => [
             'driver' => 'apc',
         ],
 
-        'array'     => [
+        'array' => [
             'driver'    => 'array',
             'serialize' => false,
         ],
 
-        'database'  => [
+        'database' => [
             'driver'          => 'database',
             'table'           => 'cache',
             'connection'      => null,
             'lock_connection' => null,
         ],
 
-        'file'      => [
+        'file' => [
             'driver' => 'file',
             'path'   => storage_path('framework/cache/data'),
         ],
@@ -59,10 +59,10 @@ return [
                 env('MEMCACHED_USERNAME'),
                 env('MEMCACHED_PASSWORD'),
             ],
-            'options'       => [
+            'options' => [
                 // Memcached::OPT_CONNECT_TIMEOUT => 2000,
             ],
-            'servers'       => [
+            'servers' => [
                 [
                     'host'   => env('MEMCACHED_HOST', '127.0.0.1'),
                     'port'   => env('MEMCACHED_PORT', 11211),
@@ -71,13 +71,13 @@ return [
             ],
         ],
 
-        'redis'     => [
+        'redis' => [
             'driver'          => 'redis',
             'connection'      => 'cache',
             'lock_connection' => 'default',
         ],
 
-        'dynamodb'  => [
+        'dynamodb' => [
             'driver'   => 'dynamodb',
             'key'      => env('AWS_ACCESS_KEY_ID'),
             'secret'   => env('AWS_SECRET_ACCESS_KEY'),
@@ -86,7 +86,7 @@ return [
             'endpoint' => env('DYNAMODB_ENDPOINT'),
         ],
 
-        'octane'    => [
+        'octane' => [
             'driver' => 'octane',
         ],
     ],
@@ -102,5 +102,5 @@ return [
     |
     */
 
-    'prefix'  => env('CACHE_PREFIX', Str::slug(env('APP_NAME', 'laravel'), '_') . '_cache_'),
+    'prefix' => env('CACHE_PREFIX', Str::slug(env('APP_NAME', 'laravel'), '_') . '_cache_'),
 ];
