@@ -11,7 +11,7 @@ return [
     |
     */
 
-    'enable'                  => env('CLOCKWORK_ENABLE', null),
+    'enable' => env('CLOCKWORK_ENABLE', null),
 
     /*
     |------------------------------------------------------------------------------------------------------------------
@@ -23,44 +23,44 @@ return [
     |
     */
 
-    'features'                => [
+    'features' => [
         // Cache usage stats and cache queries including results
-        'cache'         => [
-            'enabled'         => env('CLOCKWORK_CACHE_ENABLED', true),
+        'cache' => [
+            'enabled' => env('CLOCKWORK_CACHE_ENABLED', true),
 
             // Collect cache queries
             'collect_queries' => env('CLOCKWORK_CACHE_QUERIES', true),
 
             // Collect values from cache queries (high performance impact with a very high number of queries)
-            'collect_values'  => env('CLOCKWORK_CACHE_COLLECT_VALUES', false),
+            'collect_values' => env('CLOCKWORK_CACHE_COLLECT_VALUES', false),
         ],
 
         // Database usage stats and queries
-        'database'      => [
-            'enabled'                  => env('CLOCKWORK_DATABASE_ENABLED', true),
+        'database' => [
+            'enabled' => env('CLOCKWORK_DATABASE_ENABLED', true),
 
             // Collect database queries (high performance impact with a very high number of queries)
-            'collect_queries'          => env('CLOCKWORK_DATABASE_COLLECT_QUERIES', true),
+            'collect_queries' => env('CLOCKWORK_DATABASE_COLLECT_QUERIES', true),
 
             // Collect details of models updates (high performance impact with a lot of model updates)
-            'collect_models_actions'   => env('CLOCKWORK_DATABASE_COLLECT_MODELS_ACTIONS', true),
+            'collect_models_actions' => env('CLOCKWORK_DATABASE_COLLECT_MODELS_ACTIONS', true),
 
             // Collect details of retrieved models (very high performance impact with a lot of models retrieved)
             'collect_models_retrieved' => env('CLOCKWORK_DATABASE_COLLECT_MODELS_RETRIEVED', false),
 
             // Query execution time threshold in milliseconds after which the query will be marked as slow
-            'slow_threshold'           => env('CLOCKWORK_DATABASE_SLOW_THRESHOLD'),
+            'slow_threshold' => env('CLOCKWORK_DATABASE_SLOW_THRESHOLD'),
 
             // Collect only slow database queries
-            'slow_only'                => env('CLOCKWORK_DATABASE_SLOW_ONLY', false),
+            'slow_only' => env('CLOCKWORK_DATABASE_SLOW_ONLY', false),
 
             // Detect and report duplicate queries
             'detect_duplicate_queries' => env('CLOCKWORK_DATABASE_DETECT_DUPLICATE_QUERIES', false),
         ],
 
         // Dispatched events
-        'events'        => [
-            'enabled'        => env('CLOCKWORK_EVENTS_ENABLED', true),
+        'events' => [
+            'enabled' => env('CLOCKWORK_EVENTS_ENABLED', true),
 
             // Ignored events (framework events are ignored by default)
             'ignored_events' => [
@@ -70,7 +70,7 @@ return [
         ],
 
         // Laravel log (you can still log directly to Clockwork with laravel log disabled)
-        'log'           => [
+        'log' => [
             'enabled' => env('CLOCKWORK_LOG_ENABLED', true),
         ],
 
@@ -80,35 +80,35 @@ return [
         ],
 
         // Performance metrics
-        'performance'   => [
+        'performance' => [
             // Allow collecting of client metrics. Requires separate clockwork-browser npm package.
             'client_metrics' => env('CLOCKWORK_PERFORMANCE_CLIENT_METRICS', true),
         ],
 
         // Dispatched queue jobs
-        'queue'         => [
+        'queue' => [
             'enabled' => env('CLOCKWORK_QUEUE_ENABLED', true),
         ],
 
         // Redis commands
-        'redis'         => [
+        'redis' => [
             'enabled' => env('CLOCKWORK_REDIS_ENABLED', true),
         ],
 
         // Routes list
-        'routes'        => [
-            'enabled'         => env('CLOCKWORK_ROUTES_ENABLED', false),
+        'routes' => [
+            'enabled' => env('CLOCKWORK_ROUTES_ENABLED', false),
 
             // Collect only routes from particular namespaces (only application routes by default)
             'only_namespaces' => ['App'],
         ],
 
         // Rendered views
-        'views'         => [
-            'enabled'           => env('CLOCKWORK_VIEWS_ENABLED', true),
+        'views' => [
+            'enabled' => env('CLOCKWORK_VIEWS_ENABLED', true),
 
             // Collect views including view data (high performance impact with a high number of views)
-            'collect_data'      => env('CLOCKWORK_VIEWS_COLLECT_DATA', false),
+            'collect_data' => env('CLOCKWORK_VIEWS_COLLECT_DATA', false),
 
             // Use Twig profiler instead of Laravel events for apps using laravel-twigbridge (more precise, but does
             // not support collecting view data)
@@ -126,7 +126,7 @@ return [
     |
     */
 
-    'web'                     => env('CLOCKWORK_WEB', true),
+    'web' => env('CLOCKWORK_WEB', true),
 
     /*
     |------------------------------------------------------------------------------------------------------------------
@@ -139,7 +139,7 @@ return [
     |
     */
 
-    'toolbar'                 => env('CLOCKWORK_TOOLBAR', true),
+    'toolbar' => env('CLOCKWORK_TOOLBAR', true),
 
     /*
     |------------------------------------------------------------------------------------------------------------------
@@ -150,33 +150,33 @@ return [
     |
     */
 
-    'requests'                => [
+    'requests' => [
         // With on-demand mode enabled, Clockwork will only profile requests when the browser extension is open or you
         // manually pass a "clockwork-profile" cookie or get/post data key.
         // Optionally you can specify a "secret" that has to be passed as the value to enable profiling.
-        'on_demand'        => env('CLOCKWORK_REQUESTS_ON_DEMAND', false),
+        'on_demand' => env('CLOCKWORK_REQUESTS_ON_DEMAND', false),
 
         // Collect only errors (requests with HTTP 4xx and 5xx responses)
-        'errors_only'      => env('CLOCKWORK_REQUESTS_ERRORS_ONLY', false),
+        'errors_only' => env('CLOCKWORK_REQUESTS_ERRORS_ONLY', false),
 
         // Response time threshold in milliseconds after which the request will be marked as slow
-        'slow_threshold'   => env('CLOCKWORK_REQUESTS_SLOW_THRESHOLD'),
+        'slow_threshold' => env('CLOCKWORK_REQUESTS_SLOW_THRESHOLD'),
 
         // Collect only slow requests
-        'slow_only'        => env('CLOCKWORK_REQUESTS_SLOW_ONLY', false),
+        'slow_only' => env('CLOCKWORK_REQUESTS_SLOW_ONLY', false),
 
         // Sample the collected requests (e.g. set to 100 to collect only 1 in 100 requests)
-        'sample'           => env('CLOCKWORK_REQUESTS_SAMPLE', false),
+        'sample' => env('CLOCKWORK_REQUESTS_SAMPLE', false),
 
         // List of URIs that should not be collected
-        'except'           => [
+        'except' => [
             '/horizon/.*', // Laravel Horizon requests
             '/telescope/.*', // Laravel Telescope requests
             '/_debugbar/.*', // Laravel DebugBar requests
         ],
 
         // List of URIs that should be collected, any other URI will not be collected if not empty
-        'only'             => [
+        'only' => [
             // '/api/.*'
         ],
 
@@ -194,22 +194,22 @@ return [
     |
     */
 
-    'artisan'                 => [
+    'artisan' => [
         // Enable or disable collection of executed Artisan commands
-        'collect'                 => env('CLOCKWORK_ARTISAN_COLLECT', true),
+        'collect' => env('CLOCKWORK_ARTISAN_COLLECT', true),
 
         // List of commands that should not be collected (built-in commands are not collected by default)
-        'except'                  => [
+        'except' => [
             // 'inspire'
         ],
 
         // List of commands that should be collected, any other command will not be collected if not empty
-        'only'                    => [
+        'only' => [
             // 'inspire'
         ],
 
         // Enable or disable collection of command output
-        'collect_output'          => env('CLOCKWORK_ARTISAN_COLLECT_OUTPUT', false),
+        'collect_output' => env('CLOCKWORK_ARTISAN_COLLECT_OUTPUT', false),
 
         // Enable or disable collection of built-in Laravel commands
         'except_laravel_commands' => env('CLOCKWORK_ARTISAN_EXCEPT_LARAVEL_COMMANDS', true),
@@ -225,17 +225,17 @@ return [
     |
     */
 
-    'queue'                   => [
+    'queue' => [
         // Enable or disable collection of executed queue jobs
         'collect' => env('CLOCKWORK_QUEUE_COLLECT', true),
 
         // List of queue jobs that should not be collected
-        'except'  => [
+        'except' => [
             // App\Jobs\ExpensiveJob::class
         ],
 
         // List of queue jobs that should be collected, any other queue job will not be collected if not empty
-        'only'    => [
+        'only' => [
             // App\Jobs\BuggyJob::class
         ],
     ],
@@ -250,12 +250,12 @@ return [
     |
     */
 
-    'tests'                   => [
+    'tests' => [
         // Enable or disable collection of ran tests
         'collect' => env('CLOCKWORK_TESTS_COLLECT', false),
 
         // List of tests that should not be collected
-        'except'  => [
+        'except' => [
             // Tests\Unit\ExampleTest::class
         ],
     ],
@@ -269,7 +269,7 @@ return [
     |
     */
 
-    'collect_data_always'     => env('CLOCKWORK_COLLECT_DATA_ALWAYS', false),
+    'collect_data_always' => env('CLOCKWORK_COLLECT_DATA_ALWAYS', false),
 
     /*
     |------------------------------------------------------------------------------------------------------------------
@@ -282,22 +282,22 @@ return [
     |
     */
 
-    'storage'                 => env('CLOCKWORK_STORAGE', 'files'),
+    'storage' => env('CLOCKWORK_STORAGE', 'files'),
 
     // Path where the Clockwork metadata is stored
-    'storage_files_path'      => env('CLOCKWORK_STORAGE_FILES_PATH', storage_path('clockwork')),
+    'storage_files_path' => env('CLOCKWORK_STORAGE_FILES_PATH', storage_path('clockwork')),
 
     // Compress the metadata files using gzip, trading a little bit of performance for lower disk usage
-    'storage_files_compress'  => env('CLOCKWORK_STORAGE_FILES_COMPRESS', false),
+    'storage_files_compress' => env('CLOCKWORK_STORAGE_FILES_COMPRESS', false),
 
     // SQL database to use, can be a name of database configured in database.php or a path to a SQLite file
-    'storage_sql_database'    => env('CLOCKWORK_STORAGE_SQL_DATABASE', storage_path('clockwork.sqlite')),
+    'storage_sql_database' => env('CLOCKWORK_STORAGE_SQL_DATABASE', storage_path('clockwork.sqlite')),
 
     // SQL table name to use, the table is automatically created and updated when needed
-    'storage_sql_table'       => env('CLOCKWORK_STORAGE_SQL_TABLE', 'clockwork'),
+    'storage_sql_table' => env('CLOCKWORK_STORAGE_SQL_TABLE', 'clockwork'),
 
     // Maximum lifetime of collected metadata in minutes, older requests will automatically be deleted, false to disable
-    'storage_expiration'      => env('CLOCKWORK_STORAGE_EXPIRATION', 60 * 24 * 7),
+    'storage_expiration' => env('CLOCKWORK_STORAGE_EXPIRATION', 60 * 24 * 7),
 
     /*
     |------------------------------------------------------------------------------------------------------------------
@@ -310,7 +310,7 @@ return [
     |
     */
 
-    'authentication'          => env('CLOCKWORK_AUTHENTICATION', false),
+    'authentication' => env('CLOCKWORK_AUTHENTICATION', false),
 
     // Password for the simple authentication
     'authentication_password' => env('CLOCKWORK_AUTHENTICATION_PASSWORD', 'VerySecretPassword'),
@@ -326,15 +326,15 @@ return [
     |
     */
 
-    'stack_traces'            => [
+    'stack_traces' => [
         // Enable or disable collecting of stack traces
-        'enabled'         => env('CLOCKWORK_STACK_TRACES_ENABLED', true),
+        'enabled' => env('CLOCKWORK_STACK_TRACES_ENABLED', true),
 
         // Limit the number of frames to be collected
-        'limit'           => env('CLOCKWORK_STACK_TRACES_LIMIT', 10),
+        'limit' => env('CLOCKWORK_STACK_TRACES_LIMIT', 10),
 
         // List of vendor names to skip when determining caller, common vendors are automatically added
-        'skip_vendors'    => [
+        'skip_vendors' => [
             // 'phpunit'
         ],
 
@@ -344,7 +344,7 @@ return [
         ],
 
         // List of class names to skip when determining caller
-        'skip_classes'    => [
+        'skip_classes' => [
             // App\CustomLog::class
         ],
     ],
@@ -360,10 +360,10 @@ return [
     */
 
     // Maximum depth of serialized multi-level arrays and objects
-    'serialization_depth'     => env('CLOCKWORK_SERIALIZATION_DEPTH', 10),
+    'serialization_depth' => env('CLOCKWORK_SERIALIZATION_DEPTH', 10),
 
     // A list of classes that will never be serialized (e.g. a common service container class)
-    'serialization_blackbox'  => [
+    'serialization_blackbox' => [
         \Illuminate\Container\Container::class,
         \Illuminate\Foundation\Application::class,
         \Laravel\Lumen\Application::class,
@@ -379,7 +379,7 @@ return [
     |
     */
 
-    'register_helpers'        => env('CLOCKWORK_REGISTER_HELPERS', true),
+    'register_helpers' => env('CLOCKWORK_REGISTER_HELPERS', true),
 
     /*
     |------------------------------------------------------------------------------------------------------------------
@@ -391,7 +391,7 @@ return [
     |
     */
 
-    'headers'                 => [
+    'headers' => [
         // 'Accept' => 'application/vnd.com.whatever.v1+json',
     ],
 
@@ -407,5 +407,5 @@ return [
     |
     */
 
-    'server_timing'           => env('CLOCKWORK_SERVER_TIMING', 10),
+    'server_timing' => env('CLOCKWORK_SERVER_TIMING', 10),
 ];
