@@ -14,7 +14,7 @@ return [
     |
     */
 
-    'default'     => env('DB_CONNECTION', 'mysql'),
+    'default' => env('DB_CONNECTION', 'mysql'),
 
     /*
     |--------------------------------------------------------------------------
@@ -41,7 +41,7 @@ return [
             'foreign_key_constraints' => env('DB_FOREIGN_KEYS', true),
         ],
 
-        'mysql'  => [
+        'mysql' => [
             'driver'         => 'mysql',
             'url'            => env('DATABASE_URL'),
             'host'           => env('DB_HOST', '127.0.0.1'),
@@ -61,7 +61,7 @@ return [
             ]) : [],
         ],
 
-        'pgsql'  => [
+        'pgsql' => [
             'driver'         => 'pgsql',
             'url'            => env('DATABASE_URL'),
             'host'           => env('DB_HOST', '127.0.0.1'),
@@ -103,7 +103,7 @@ return [
     |
     */
 
-    'migrations'  => 'migrations',
+    'migrations' => 'migrations',
 
     /*
     |--------------------------------------------------------------------------
@@ -116,8 +116,8 @@ return [
     |
     */
 
-    'redis'       => [
-        'client'  => env('REDIS_CLIENT', 'phpredis'),
+    'redis' => [
+        'client' => env('REDIS_CLIENT', 'phpredis'),
 
         'options' => [
             'cluster' => env('REDIS_CLUSTER', 'redis'),
@@ -133,7 +133,7 @@ return [
             'database' => env('REDIS_DB', '0'),
         ],
 
-        'cache'   => [
+        'cache' => [
             'url'      => env('REDIS_URL'),
             'host'     => env('REDIS_HOST', '127.0.0.1'),
             'username' => env('REDIS_USERNAME'),
