@@ -42,7 +42,7 @@ return [
     |
     */
 
-    'argon'  => [
+    'argon' => [
         'memory'  => 65536,
         'threads' => 1,
         'time'    => 4,
