@@ -3,6 +3,7 @@
 use Monolog\Handler\NullHandler;
 use Monolog\Handler\StreamHandler;
 use Monolog\Handler\SyslogUdpHandler;
+use Monolog\Processor\PsrLogMessageProcessor;
 use Rollbar\Laravel\MonologHandler;
 
 return [
@@ -17,7 +18,7 @@ return [
     |
     */
 
-    'default'      => env('LOG_CHANNEL', 'stack'),
+    'default' => env('LOG_CHANNEL', 'stack'),
 
     /*
     |--------------------------------------------------------------------------
@@ -50,32 +51,35 @@ return [
     |
     */
 
-    'channels'     => [
-        'stack'      => [
+    'channels' => [
+        'stack' => [
             'driver'            => 'stack',
             'channels'          => ['rollbar', 'single'],
             'ignore_exceptions' => false,
         ],
 
-        'single'     => [
-            'driver' => 'single',
-            'path'   => storage_path('logs/laravel.log'),
-            'level'  => env('LOG_LEVEL', 'debug'),
+        'single' => [
+            'driver'               => 'single',
+            'path'                 => storage_path('logs/laravel.log'),
+            'level'                => env('LOG_LEVEL', 'debug'),
+            'replace_placeholders' => true,
         ],
 
-        'daily'      => [
-            'driver' => 'daily',
-            'path'   => storage_path('logs/laravel.log'),
-            'level'  => env('LOG_LEVEL', 'debug'),
-            'days'   => 14,
+        'daily' => [
+            'driver'               => 'daily',
+            'path'                 => storage_path('logs/laravel.log'),
+            'level'                => env('LOG_LEVEL', 'debug'),
+            'days'                 => 14,
+            'replace_placeholders' => true,
         ],
 
-        'slack'      => [
-            'driver'   => 'slack',
-            'url'      => env('LOG_SLACK_WEBHOOK_URL'),
-            'username' => 'Laravel Log',
-            'emoji'    => ':boom:',
-            'level'    => env('LOG_LEVEL', 'critical'),
+        'slack' => [
+            'driver'               => 'slack',
+            'url'                  => env('LOG_SLACK_WEBHOOK_URL'),
+            'username'             => 'Laravel Log',
+            'emoji'                => ':boom:',
+            'level'                => env('LOG_LEVEL', 'critical'),
+            'replace_placeholders' => true,
         ],
 
         'papertrail' => [
@@ -87,9 +91,10 @@ return [
                 'port'             => env('PAPERTRAIL_PORT'),
                 'connectionString' => 'tls://' . env('PAPERTRAIL_URL') . ':' . env('PAPERTRAIL_PORT'),
             ],
+            'processors' => [PsrLogMessageProcessor::class],
         ],
 
-        'stderr'     => [
+        'stderr' => [
             'driver'    => 'monolog',
             'level'     => env('LOG_LEVEL', 'debug'),
             'handler'   => StreamHandler::class,
@@ -97,32 +102,36 @@ return [
             'with'      => [
                 'stream' => 'php://stderr',
             ],
+            'processors' => [PsrLogMessageProcessor::class],
         ],
 
-        'syslog'     => [
-            'driver' => 'syslog',
-            'level'  => env('LOG_LEVEL', 'debug'),
+        'syslog' => [
+            'driver'               => 'syslog',
+            'level'                => env('LOG_LEVEL', 'debug'),
+            'facility'             => LOG_USER,
+            'replace_placeholders' => true,
         ],
 
-        'errorlog'   => [
-            'driver' => 'errorlog',
-            'level'  => env('LOG_LEVEL', 'debug'),
+        'errorlog' => [
+            'driver'               => 'errorlog',
+            'level'                => env('LOG_LEVEL', 'debug'),
+            'replace_placeholders' => true,
         ],
 
-        'null'       => [
+        'null' => [
             'driver'  => 'monolog',
             'handler' => NullHandler::class,
         ],
 
-        'emergency'  => [
+        'emergency' => [
             'path' => storage_path('logs/laravel.log'),
         ],
 
-        'rollbar'    => [
+        'rollbar' => [
             'driver'       => 'monolog',
             'handler'      => MonologHandler::class,
             'access_token' => env('ROLLBAR_TOKEN'),
-            'level'        => env('ROLLBAR_LEVEL', 'error'),
+            'level'        => env('ROLLBAR_LEVEL', 'debug'),
             'capture_ip'   => 'anonymize',
         ],
     ],
