@@ -12,7 +12,7 @@ return [
     |
     */
 
-    'default'  => env('MAIL_MAILER', 'smtp'),
+    'default' => env('MAIL_MAILER', 'smtp'),
 
     /*
     |--------------------------------------------------------------------------
@@ -27,13 +27,13 @@ return [
     | sending an e-mail. You will specify which one you are using for your
     | mailers below. You are free to add additional mailers as required.
     |
-    | Supported: "smtp", "sendmail", "mailgun", "ses",
+    | Supported: "smtp", "sendmail", "mailgun", "ses", "ses-v2",
     |            "postmark", "log", "array", "failover"
     |
     */
 
-    'mailers'  => [
-        'smtp'     => [
+    'mailers' => [
+        'smtp' => [
             'transport'    => 'smtp',
             'host'         => env('MAIL_HOST', 'smtp.mailgun.org'),
             'port'         => env('MAIL_PORT', 587),
@@ -44,16 +44,22 @@ return [
             'local_domain' => env('MAIL_EHLO_DOMAIN'),
         ],
 
-        'ses'      => [
+        'ses' => [
             'transport' => 'ses',
         ],
 
-        'mailgun'  => [
+        'mailgun' => [
             'transport' => 'mailgun',
+            // 'client' => [
+            //     'timeout' => 5,
+            // ],
         ],
 
         'postmark' => [
             'transport' => 'postmark',
+            // 'client' => [
+            //     'timeout' => 5,
+            // ],
         ],
 
         'sendmail' => [
@@ -61,12 +67,12 @@ return [
             'path'      => env('MAIL_SENDMAIL_PATH', '/usr/sbin/sendmail -bs -i'),
         ],
 
-        'log'      => [
+        'log' => [
             'transport' => 'log',
             'channel'   => env('MAIL_LOG_CHANNEL'),
         ],
 
-        'array'    => [
+        'array' => [
             'transport' => 'array',
         ],
 
@@ -90,7 +96,7 @@ return [
     |
     */
 
-    'from'     => [
+    'from' => [
         'address' => env('MAIL_FROM_ADDRESS', 'hello@example.com'),
         'name'    => env('MAIL_FROM_NAME', 'Example'),
     ],
