@@ -14,7 +14,7 @@ return [
     |
     */
 
-    'stateful'   => explode(',', env('SANCTUM_STATEFUL_DOMAINS', sprintf(
+    'stateful' => explode(',', env('SANCTUM_STATEFUL_DOMAINS', sprintf(
         '%s%s',
         'localhost,localhost:3000,127.0.0.1,127.0.0.1:8000,::1',
         Sanctum::currentApplicationUrlWithPort()
@@ -32,7 +32,7 @@ return [
     |
     */
 
-    'guard'      => ['web'],
+    'guard' => ['web'],
 
     /*
     |--------------------------------------------------------------------------
