@@ -9,7 +9,7 @@ return [
      * Note: when server is empty string, it will not add to response header
      */
 
-    'server'                            => '',
+    'server' => '',
 
     /*
      * X-Content-Type-Options
@@ -19,7 +19,7 @@ return [
      * Available Value: 'nosniff'
      */
 
-    'x-content-type-options'            => 'nosniff',
+    'x-content-type-options' => 'nosniff',
 
     /*
      * X-Download-Options
@@ -29,7 +29,7 @@ return [
      * Available Value: 'noopen'
      */
 
-    'x-download-options'                => 'noopen',
+    'x-download-options' => 'noopen',
 
     /*
      * X-Frame-Options
@@ -39,7 +39,7 @@ return [
      * Available Value: 'deny', 'sameorigin', 'allow-from <uri>'
      */
 
-    'x-frame-options'                   => 'deny',
+    'x-frame-options' => 'deny',
 
     /*
      * X-Permitted-Cross-Domain-Policies
@@ -62,7 +62,7 @@ return [
      * Reference: https://github.com/bepsvpt/secure-headers/issues/58#issuecomment-782332442
      */
 
-    'x-powered-by'                      => '',
+    'x-powered-by' => '',
 
     /*
      * X-XSS-Protection
@@ -72,7 +72,7 @@ return [
      * Available Value: '1', '0', '1; mode=block'
      */
 
-    'x-xss-protection'                  => '1; mode=block',
+    'x-xss-protection' => '1; mode=block',
 
     /*
      * Referrer-Policy
@@ -83,7 +83,34 @@ return [
      *                  'same-origin', 'strict-origin', 'strict-origin-when-cross-origin', 'unsafe-url'
      */
 
-    'referrer-policy'                   => 'no-referrer',
+    'referrer-policy' => 'no-referrer',
+
+    /*
+     * Cross-Origin-Embedder-Policy
+     *
+     * Reference: https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Cross-Origin-Embedder-Policy
+     *
+     * Available Value: 'unsafe-none', 'require-corp'
+     */
+    'cross-origin-embedder-policy' => 'unsafe-none',
+
+    /*
+     * Cross-Origin-Opener-Policy
+     *
+     * Reference: https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Cross-Origin-Opener-Policy
+     *
+     * Available Value: 'unsafe-none', 'same-origin-allow-popups', 'same-origin'
+     */
+    'cross-origin-opener-policy' => 'unsafe-none',
+
+    /*
+     * Cross-Origin-Resource-Policy
+     *
+     * Reference: https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Cross-Origin-Resource-Policy
+     *
+     * Available Value: 'same-site', 'same-origin', 'cross-origin'
+     */
+    'cross-origin-resource-policy' => 'cross-origin',
 
     /*
      * Clear-Site-Data
@@ -91,16 +118,16 @@ return [
      * Reference: https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Clear-Site-Data
      */
 
-    'clear-site-data'                   => [
-        'enable'            => false,
+    'clear-site-data' => [
+        'enable' => false,
 
-        'all'               => false,
+        'all' => false,
 
-        'cache'             => true,
+        'cache' => true,
 
-        'cookies'           => true,
+        'cookies' => true,
 
-        'storage'           => true,
+        'storage' => true,
 
         'executionContexts' => true,
     ],
@@ -113,14 +140,14 @@ return [
      * Please ensure your website had set up ssl/tls before enable hsts.
      */
 
-    'hsts'                              => [
-        'enable'              => false,
+    'hsts' => [
+        'enable' => false,
 
-        'max-age'             => 31536000,
+        'max-age' => 31536000,
 
         'include-sub-domains' => false,
 
-        'preload'             => false,
+        'preload' => false,
     ],
 
     /*
@@ -129,12 +156,12 @@ return [
      * Reference: https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Expect-CT
      */
 
-    'expect-ct'                         => [
-        'enable'     => false,
+    'expect-ct' => [
+        'enable' => false,
 
-        'max-age'    => 2147483648,
+        'max-age' => 2147483648,
 
-        'enforce'    => false,
+        'enforce' => false,
 
         // report uri must be absolute-URI
         'report-uri' => null,
@@ -146,214 +173,292 @@ return [
      * Reference: https://w3c.github.io/webappsec-permissions-policy/
      */
 
-    'permissions-policy'                => [
-        'enable'                          => true,
+    'permissions-policy' => [
+        'enable' => true,
 
         // https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Feature-Policy/accelerometer
-        'accelerometer'                   => [
-            'none'    => false,
-            '*'       => false,
-            'self'    => true,
+        'accelerometer' => [
+            'none' => false,
+
+            '*' => false,
+
+            'self' => true,
+
             'origins' => [],
         ],
 
         // https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Feature-Policy/ambient-light-sensor
-        'ambient-light-sensor'            => [
-            'none'    => false,
-            '*'       => false,
-            'self'    => true,
+        'ambient-light-sensor' => [
+            'none' => false,
+
+            '*' => false,
+
+            'self' => true,
+
             'origins' => [],
         ],
 
         // https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Feature-Policy/autoplay
-        'autoplay'                        => [
-            'none'    => false,
-            '*'       => false,
-            'self'    => true,
+        'autoplay' => [
+            'none' => false,
+
+            '*' => false,
+
+            'self' => true,
+
             'origins' => [],
         ],
 
         // https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Feature-Policy/battery
-        'battery'                         => [
-            'none'    => false,
-            '*'       => false,
-            'self'    => true,
+        'battery' => [
+            'none' => false,
+
+            '*' => false,
+
+            'self' => true,
+
             'origins' => [],
         ],
 
         // https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Feature-Policy/camera
-        'camera'                          => [
-            'none'    => false,
-            '*'       => false,
-            'self'    => true,
+        'camera' => [
+            'none' => false,
+
+            '*' => false,
+
+            'self' => true,
+
             'origins' => [],
         ],
 
         // https://www.chromestatus.com/feature/5690888397258752
-        'cross-origin-isolated'           => [
-            'none'    => false,
-            '*'       => false,
-            'self'    => true,
+        'cross-origin-isolated' => [
+            'none' => false,
+
+            '*' => false,
+
+            'self' => true,
+
             'origins' => [],
         ],
 
         // https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Feature-Policy/display-capture
-        'display-capture'                 => [
-            'none'    => false,
-            '*'       => false,
-            'self'    => true,
+        'display-capture' => [
+            'none' => false,
+
+            '*' => false,
+
+            'self' => true,
+
             'origins' => [],
         ],
 
         // https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Feature-Policy/document-domain
-        'document-domain'                 => [
-            'none'    => false,
-            '*'       => true,
-            'self'    => false,
+        'document-domain' => [
+            'none' => false,
+
+            '*' => true,
+
+            'self' => false,
+
             'origins' => [],
         ],
 
         // https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Feature-Policy/encrypted-media
-        'encrypted-media'                 => [
-            'none'    => false,
-            '*'       => false,
-            'self'    => true,
+        'encrypted-media' => [
+            'none' => false,
+
+            '*' => false,
+
+            'self' => true,
+
             'origins' => [],
         ],
 
         // https://wicg.github.io/page-lifecycle/#execution-while-not-rendered
-        'execution-while-not-rendered'    => [
-            'none'    => false,
-            '*'       => true,
-            'self'    => false,
+        'execution-while-not-rendered' => [
+            'none' => false,
+
+            '*' => true,
+
+            'self' => false,
+
             'origins' => [],
         ],
 
         // https://wicg.github.io/page-lifecycle/#execution-while-out-of-viewport
         'execution-while-out-of-viewport' => [
-            'none'    => false,
-            '*'       => true,
-            'self'    => false,
+            'none' => false,
+
+            '*' => true,
+
+            'self' => false,
+
             'origins' => [],
         ],
 
         // https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Feature-Policy/fullscreen
-        'fullscreen'                      => [
-            'none'    => false,
-            '*'       => false,
-            'self'    => true,
+        'fullscreen' => [
+            'none' => false,
+
+            '*' => false,
+
+            'self' => true,
+
             'origins' => [],
         ],
 
         // https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Feature-Policy/geolocation
-        'geolocation'                     => [
-            'none'    => false,
-            '*'       => false,
-            'self'    => true,
+        'geolocation' => [
+            'none' => false,
+
+            '*' => false,
+
+            'self' => true,
+
             'origins' => [],
         ],
 
         // https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Feature-Policy/gyroscope
-        'gyroscope'                       => [
-            'none'    => false,
-            '*'       => false,
-            'self'    => true,
+        'gyroscope' => [
+            'none' => false,
+
+            '*' => false,
+
+            'self' => true,
+
             'origins' => [],
         ],
 
         // https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Feature-Policy/magnetometer
-        'magnetometer'                    => [
-            'none'    => false,
-            '*'       => false,
-            'self'    => true,
+        'magnetometer' => [
+            'none' => false,
+
+            '*' => false,
+
+            'self' => true,
+
             'origins' => [],
         ],
 
         // https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Feature-Policy/microphone
-        'microphone'                      => [
-            'none'    => false,
-            '*'       => false,
-            'self'    => true,
+        'microphone' => [
+            'none' => false,
+
+            '*' => false,
+
+            'self' => true,
+
             'origins' => [],
         ],
 
         // https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Feature-Policy/midi
-        'midi'                            => [
-            'none'    => false,
-            '*'       => false,
-            'self'    => true,
+        'midi' => [
+            'none' => false,
+
+            '*' => false,
+
+            'self' => true,
+
             'origins' => [],
         ],
 
         // https://drafts.csswg.org/css-nav-1/
-        'navigation-override'             => [
-            'none'    => false,
-            '*'       => false,
-            'self'    => true,
+        'navigation-override' => [
+            'none' => false,
+
+            '*' => false,
+
+            'self' => true,
+
             'origins' => [],
         ],
 
         // https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Feature-Policy/payment
-        'payment'                         => [
-            'none'    => false,
-            '*'       => false,
-            'self'    => true,
+        'payment' => [
+            'none' => false,
+
+            '*' => false,
+
+            'self' => true,
+
             'origins' => [],
         ],
 
         // https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Feature-Policy/picture-in-picture
-        'picture-in-picture'              => [
-            'none'    => false,
-            '*'       => true,
-            'self'    => false,
+        'picture-in-picture' => [
+            'none' => false,
+
+            '*' => true,
+
+            'self' => false,
+
             'origins' => [],
         ],
 
         // https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Feature-Policy/publickey-credentials-get
-        'publickey-credentials-get'       => [
-            'none'    => false,
-            '*'       => false,
-            'self'    => true,
+        'publickey-credentials-get' => [
+            'none' => false,
+
+            '*' => false,
+
+            'self' => true,
+
             'origins' => [],
         ],
 
         // https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Feature-Policy/screen-wake-lock
-        'screen-wake-lock'                => [
-            'none'    => false,
-            '*'       => false,
-            'self'    => true,
+        'screen-wake-lock' => [
+            'none' => false,
+
+            '*' => false,
+
+            'self' => true,
+
             'origins' => [],
         ],
 
         // https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Feature-Policy/sync-xhr
-        'sync-xhr'                        => [
-            'none'    => false,
-            '*'       => true,
-            'self'    => false,
+        'sync-xhr' => [
+            'none' => false,
+
+            '*' => true,
+
+            'self' => false,
+
             'origins' => [],
         ],
 
         // https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Feature-Policy/usb
-        'usb'                             => [
-            'none'    => false,
-            '*'       => false,
-            'self'    => true,
+        'usb' => [
+            'none' => false,
+
+            '*' => false,
+
+            'self' => true,
+
             'origins' => [],
         ],
 
         // https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Feature-Policy/web-share
-        'web-share'                       => [
-            'none'    => false,
-            '*'       => false,
-            'self'    => true,
+        'web-share' => [
+            'none' => false,
+
+            '*' => false,
+
+            'self' => true,
+
             'origins' => [],
         ],
 
         // https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Feature-Policy/xr-spatial-tracking
-        'xr-spatial-tracking'             => [
-            'none'    => false,
-            '*'       => false,
-            'self'    => true,
+        'xr-spatial-tracking' => [
+            'none' => false,
+
+            '*' => false,
+
+            'self' => true,
+
             'origins' => [],
         ],
     ],
@@ -364,36 +469,36 @@ return [
      * Reference: https://developer.mozilla.org/en-US/docs/Web/HTTP/CSP
      */
 
-    'csp'                               => [
-        'enable'                    => true,
+    'csp' => [
+        'enable' => true,
 
         // https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Content-Security-Policy-Report-Only
-        'report-only'               => false,
+        'report-only' => false,
 
         // https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Content-Security-Policy/report-to
-        'report-to'                 => '',
+        'report-to' => '',
 
         // https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Content-Security-Policy/report-uri
-        'report-uri'                => [
+        'report-uri' => [
             // uri
         ],
 
         // https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Content-Security-Policy/block-all-mixed-content
-        'block-all-mixed-content'   => false,
+        'block-all-mixed-content' => false,
 
         // https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Content-Security-Policy/upgrade-insecure-requests
         'upgrade-insecure-requests' => false,
 
         // https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Content-Security-Policy/base-uri
-        'base-uri'                  => [
+        'base-uri' => [
         ],
 
         // https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Content-Security-Policy/child-src
-        'child-src'                 => [
+        'child-src' => [
         ],
 
         // https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Content-Security-Policy/connect-src
-        'connect-src'               => [
+        'connect-src' => [
             'self'  => true,
             'allow' => [
                 'carbonaut.dev',
@@ -402,13 +507,13 @@ return [
         ],
 
         // https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Content-Security-Policy/default-src
-        'default-src'               => [
+        'default-src' => [
             'self' => true,
         ],
 
         // https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Content-Security-Policy/font-src
-        'font-src'                  => [
-            'allow'   => [
+        'font-src' => [
+            'allow' => [
                 '*.typekit.net',
                 '*.gstatic.com',
             ],
@@ -418,19 +523,19 @@ return [
         ],
 
         // https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Content-Security-Policy/form-action
-        'form-action'               => [
+        'form-action' => [
         ],
 
         // https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Content-Security-Policy/frame-ancestors
-        'frame-ancestors'           => [
+        'frame-ancestors' => [
         ],
 
         // https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Content-Security-Policy/frame-src
-        'frame-src'                 => [
+        'frame-src' => [
         ],
 
         // https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Content-Security-Policy/img-src
-        'img-src'                   => [
+        'img-src' => [
             'self'    => true,
             'schemes' => [
                 'data:',
@@ -438,29 +543,29 @@ return [
         ],
 
         // https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Content-Security-Policy/manifest-src
-        'manifest-src'              => [
+        'manifest-src' => [
         ],
 
         // https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Content-Security-Policy/media-src
-        'media-src'                 => [
+        'media-src' => [
         ],
 
         // https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Content-Security-Policy/navigate-to
-        'navigate-to'               => [
+        'navigate-to' => [
             'unsafe-allow-redirects' => false,
         ],
 
         // https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Content-Security-Policy/object-src
-        'object-src'                => [
+        'object-src' => [
         ],
 
         // https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Content-Security-Policy/plugin-types
-        'plugin-types'              => [
+        'plugin-types' => [
             // 'application/pdf',
         ],
 
         // https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Content-Security-Policy/prefetch-src
-        'prefetch-src'              => [
+        'prefetch-src' => [
         ],
 
         // https://w3c.github.io/webappsec-trusted-types/dist/spec/#integration-with-content-security-policy
@@ -469,68 +574,68 @@ return [
         ],
 
         // https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Content-Security-Policy/sandbox
-        'sandbox'                   => [
-            'enable'                                  => false,
+        'sandbox' => [
+            'enable' => false,
 
             'allow-downloads-without-user-activation' => false,
 
-            'allow-forms'                             => false,
+            'allow-forms' => false,
 
-            'allow-modals'                            => false,
+            'allow-modals' => false,
 
-            'allow-orientation-lock'                  => false,
+            'allow-orientation-lock' => false,
 
-            'allow-pointer-lock'                      => false,
+            'allow-pointer-lock' => false,
 
-            'allow-popups'                            => false,
+            'allow-popups' => false,
 
-            'allow-popups-to-escape-sandbox'          => false,
+            'allow-popups-to-escape-sandbox' => false,
 
-            'allow-presentation'                      => false,
+            'allow-presentation' => false,
 
-            'allow-same-origin'                       => false,
+            'allow-same-origin' => false,
 
-            'allow-scripts'                           => false,
+            'allow-scripts' => false,
 
             'allow-storage-access-by-user-activation' => false,
 
-            'allow-top-navigation'                    => false,
+            'allow-top-navigation' => false,
 
             'allow-top-navigation-by-user-activation' => false,
         ],
 
         // https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Content-Security-Policy/script-src
-        'script-src'                => [
-            'none'           => false,
+        'script-src' => [
+            'none' => false,
 
-            'self'           => true,
+            'self' => true,
 
-            'report-sample'  => false,
+            'report-sample' => false,
 
-            'allow'          => [
+            'allow' => [
                 // 'url',
             ],
 
-            'schemes'        => [
+            'schemes' => [
                 // 'data:',
                 // 'https:',
             ],
 
             // followings are only work for `script` and `style` related directives
 
-            'unsafe-inline'  => true,
+            'unsafe-inline' => true,
 
-            'unsafe-eval'    => true,
+            'unsafe-eval' => true,
 
             // https://www.w3.org/TR/CSP3/#unsafe-hashes-usage
-            'unsafe-hashes'  => false,
+            'unsafe-hashes' => false,
 
             // Enable `strict-dynamic` will *ignore* `self`, `unsafe-inline`,
             // `allow` and `schemes`. You can find more information from:
             // https://www.w3.org/TR/CSP3/#strict-dynamic-usage
             'strict-dynamic' => false,
 
-            'hashes'         => [
+            'hashes' => [
                 'sha256' => [
                     // 'sha256-hash-value-with-base64-encode',
                 ],
@@ -546,15 +651,15 @@ return [
         ],
 
         // https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Content-Security-Policy/script-src-attr
-        'script-src-attr'           => [
+        'script-src-attr' => [
         ],
 
         // https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Content-Security-Policy/script-src-elem
-        'script-src-elem'           => [
+        'script-src-elem' => [
         ],
 
         // https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Content-Security-Policy/style-src
-        'style-src'                 => [
+        'style-src' => [
             'unsafe-inline' => true,
             'self'          => true,
             'allow'         => [
@@ -563,27 +668,27 @@ return [
         ],
 
         // https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Content-Security-Policy/style-src-attr
-        'style-src-attr'            => [
+        'style-src-attr' => [
         ],
 
         // https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Content-Security-Policy/style-src-elem
-        'style-src-elem'            => [
+        'style-src-elem' => [
         ],
 
         // https://w3c.github.io/webappsec-trusted-types/dist/spec/#trusted-types-csp-directive
-        'trusted-types'             => [
-            'enable'           => false,
+        'trusted-types' => [
+            'enable' => false,
 
             'allow-duplicates' => false,
 
-            'default'          => false,
+            'default' => false,
 
-            'policies'         => [
+            'policies' => [
             ],
         ],
 
         // https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Content-Security-Policy/worker-src
-        'worker-src'                => [
+        'worker-src' => [
         ],
     ],
 ];
