@@ -13,7 +13,7 @@ return [
     |
     */
 
-    'mailgun'  => [
+    'mailgun' => [
         'domain'   => env('MAILGUN_DOMAIN'),
         'secret'   => env('MAILGUN_SECRET'),
         'endpoint' => env('MAILGUN_ENDPOINT', 'api.mailgun.net'),
@@ -24,7 +24,7 @@ return [
         'token' => env('POSTMARK_TOKEN'),
     ],
 
-    'ses'      => [
+    'ses' => [
         'key'    => env('AWS_ACCESS_KEY_ID'),
         'secret' => env('AWS_SECRET_ACCESS_KEY'),
         'region' => env('AWS_DEFAULT_REGION', 'us-east-1'),
