@@ -17,7 +17,7 @@ return [
     |
     */
 
-    'driver'          => env('SESSION_DRIVER', 'file'),
+    'driver' => env('SESSION_DRIVER', 'file'),
 
     /*
     |--------------------------------------------------------------------------
@@ -30,7 +30,7 @@ return [
     |
     */
 
-    'lifetime'        => env('SESSION_LIFETIME', 120),
+    'lifetime' => env('SESSION_LIFETIME', 120),
 
     'expire_on_close' => false,
 
@@ -45,7 +45,7 @@ return [
     |
     */
 
-    'encrypt'         => false,
+    'encrypt' => false,
 
     /*
     |--------------------------------------------------------------------------
@@ -58,7 +58,7 @@ return [
     |
     */
 
-    'files'           => storage_path('framework/sessions'),
+    'files' => storage_path('framework/sessions'),
 
     /*
     |--------------------------------------------------------------------------
@@ -71,7 +71,7 @@ return [
     |
     */
 
-    'connection'      => env('SESSION_CONNECTION'),
+    'connection' => env('SESSION_CONNECTION'),
 
     /*
     |--------------------------------------------------------------------------
@@ -84,7 +84,7 @@ return [
     |
     */
 
-    'table'           => 'sessions',
+    'table' => 'sessions',
 
     /*
     |--------------------------------------------------------------------------
@@ -99,7 +99,7 @@ return [
     |
     */
 
-    'store'           => env('SESSION_STORE'),
+    'store' => env('SESSION_STORE'),
 
     /*
     |--------------------------------------------------------------------------
@@ -112,7 +112,7 @@ return [
     |
     */
 
-    'lottery'         => [2, 100],
+    'lottery' => [2, 100],
 
     /*
     |--------------------------------------------------------------------------
@@ -125,7 +125,7 @@ return [
     |
     */
 
-    'cookie'          => env(
+    'cookie' => env(
         'SESSION_COOKIE',
         Str::slug(env('APP_NAME', 'laravel'), '_') . '_session'
     ),
@@ -141,7 +141,7 @@ return [
     |
     */
 
-    'path'            => '/',
+    'path' => '/',
 
     /*
     |--------------------------------------------------------------------------
@@ -154,7 +154,7 @@ return [
     |
     */
 
-    'domain'          => env('SESSION_DOMAIN'),
+    'domain' => env('SESSION_DOMAIN'),
 
     /*
     |--------------------------------------------------------------------------
@@ -167,7 +167,7 @@ return [
     |
     */
 
-    'secure'          => env('SESSION_SECURE_COOKIE'),
+    'secure' => env('SESSION_SECURE_COOKIE'),
 
     /*
     |--------------------------------------------------------------------------
@@ -180,7 +180,7 @@ return [
     |
     */
 
-    'http_only'       => true,
+    'http_only' => true,
 
     /*
     |--------------------------------------------------------------------------
@@ -195,5 +195,5 @@ return [
     |
     */
 
-    'same_site'       => 'lax',
+    'same_site' => 'lax',
 ];
