@@ -13,7 +13,7 @@ return [
      * This is the model used by the Db Translation loader. You can put any model here
      * that extends Spatie\TranslationLoader\LanguageLine.
      */
-    'model'               => App\Models\LanguageLine::class,
+    'model' => App\Models\LanguageLine::class,
 
     // This is the translation manager which overrides the default Laravel `translation.loader`
     'translation_manager' => Spatie\TranslationLoader\TranslationLoaderManager::class,
