@@ -12,7 +12,7 @@ return [
     |
     */
 
-    'paths'    => [
+    'paths' => [
         resource_path('views'),
     ],
 
