@@ -2,6 +2,8 @@
 
 namespace Database\Factories;
 
+use App\Enums\Language;
+use App\Enums\LanguageLineGroup;
 use Illuminate\Database\Eloquent\Factories\Factory;
 
 /**
@@ -14,8 +16,30 @@ class LanguageLineFactory extends Factory
      *
      * @return array<string, mixed>
      */
-    public function definition()
+    public function definition(): array
     {
-        return [];
+        return [
+            'group' => LanguageLineGroup::randomCase(),
+            'key'   => fake()->word(),
+            'text'  => [
+                Language::randomCase()->value => fake()->sentence(),
+            ],
+        ];
+    }
+
+    /**
+     * Indicate that the language line's text will translate to the given locale.
+     *
+     * @param string $locale
+     *
+     * @return static
+     */
+    public function withLocale(string $locale): static
+    {
+        return $this->state(fn (array $attributes) => [
+            'text' => [
+                $locale => fake()->sentence(),
+            ],
+        ]);
     }
 }
