@@ -1,6 +1,5 @@
 <?php
 
-use App\Enums\Language;
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
@@ -8,8 +7,6 @@ use Illuminate\Support\Facades\Schema;
 return new class() extends Migration {
     /**
      * Run the migrations.
-     *
-     * @return void
      */
     public function up(): void
     {
@@ -20,7 +17,7 @@ return new class() extends Migration {
             $table->timestamp('email_verified_at')->nullable();
             $table->string('password');
             $table->rememberToken();
-            $table->string('language')->default(Language::English->value);
+            $table->string('language')->default('en');
             $table->integer('email_verification_code')->nullable();
             $table->datetime('email_verification_code_expires_at')->nullable();
             $table->timestamps();
@@ -29,8 +26,6 @@ return new class() extends Migration {
 
     /**
      * Reverse the migrations.
-     *
-     * @return void
      */
     public function down(): void
     {
