@@ -4,8 +4,7 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-class CreateDevicesTable extends Migration
-{
+return new class() extends Migration {
     /**
      * Run the migrations.
      *
@@ -45,4 +44,4 @@ class CreateDevicesTable extends Migration
     {
         Schema::dropIfExists('devices');
     }
-}
+};
