@@ -12,98 +12,98 @@ return [
     |
     */
 
-    'accepted'             => 'The :attribute must be accepted.',
-    'accepted_if'          => 'The :attribute must be accepted when :other is :value.',
-    'active_url'           => 'The :attribute is not a valid URL.',
-    'after'                => 'The :attribute must be a date after :date.',
-    'after_or_equal'       => 'The :attribute must be a date after or equal to :date.',
-    'alpha'                => 'The :attribute must only contain letters.',
-    'alpha_dash'           => 'The :attribute must only contain letters, numbers, dashes and underscores.',
-    'alpha_num'            => 'The :attribute must only contain letters and numbers.',
-    'array'                => 'The :attribute must be an array.',
-    'before'               => 'The :attribute must be a date before :date.',
-    'before_or_equal'      => 'The :attribute must be a date before or equal to :date.',
-    'between'              => [
+    'accepted'        => 'The :attribute must be accepted.',
+    'accepted_if'     => 'The :attribute must be accepted when :other is :value.',
+    'active_url'      => 'The :attribute is not a valid URL.',
+    'after'           => 'The :attribute must be a date after :date.',
+    'after_or_equal'  => 'The :attribute must be a date after or equal to :date.',
+    'alpha'           => 'The :attribute must only contain letters.',
+    'alpha_dash'      => 'The :attribute must only contain letters, numbers, dashes and underscores.',
+    'alpha_num'       => 'The :attribute must only contain letters and numbers.',
+    'array'           => 'The :attribute must be an array.',
+    'before'          => 'The :attribute must be a date before :date.',
+    'before_or_equal' => 'The :attribute must be a date before or equal to :date.',
+    'between'         => [
         'array'   => 'The :attribute must have between :min and :max items.',
         'file'    => 'The :attribute must be between :min and :max kilobytes.',
         'numeric' => 'The :attribute must be between :min and :max.',
         'string'  => 'The :attribute must be between :min and :max characters.',
     ],
-    'boolean'              => 'The :attribute field must be true or false.',
-    'confirmed'            => 'The :attribute confirmation does not match.',
-    'current_password'     => 'The password is incorrect.',
-    'date'                 => 'The :attribute is not a valid date.',
-    'date_equals'          => 'The :attribute must be a date equal to :date.',
-    'date_format'          => 'The :attribute does not match the format :format.',
-    'declined'             => 'The :attribute must be declined.',
-    'declined_if'          => 'The :attribute must be declined when :other is :value.',
-    'different'            => 'The :attribute and :other must be different.',
-    'digits'               => 'The :attribute must be :digits digits.',
-    'digits_between'       => 'The :attribute must be between :min and :max digits.',
-    'dimensions'           => 'The :attribute has invalid image dimensions.',
-    'distinct'             => 'The :attribute field has a duplicate value.',
-    'doesnt_end_with'      => 'The :attribute may not end with one of the following: :values.',
-    'doesnt_start_with'    => 'The :attribute may not start with one of the following: :values.',
-    'email'                => 'The :attribute must be a valid email address.',
-    'ends_with'            => 'The :attribute must end with one of the following: :values.',
-    'enum'                 => 'The selected :attribute is invalid.',
-    'exists'               => 'The selected :attribute is invalid.',
-    'file'                 => 'The :attribute must be a file.',
-    'filled'               => 'The :attribute field must have a value.',
-    'gt'                   => [
+    'boolean'           => 'The :attribute field must be true or false.',
+    'confirmed'         => 'The :attribute confirmation does not match.',
+    'current_password'  => 'The password is incorrect.',
+    'date'              => 'The :attribute is not a valid date.',
+    'date_equals'       => 'The :attribute must be a date equal to :date.',
+    'date_format'       => 'The :attribute does not match the format :format.',
+    'declined'          => 'The :attribute must be declined.',
+    'declined_if'       => 'The :attribute must be declined when :other is :value.',
+    'different'         => 'The :attribute and :other must be different.',
+    'digits'            => 'The :attribute must be :digits digits.',
+    'digits_between'    => 'The :attribute must be between :min and :max digits.',
+    'dimensions'        => 'The :attribute has invalid image dimensions.',
+    'distinct'          => 'The :attribute field has a duplicate value.',
+    'doesnt_end_with'   => 'The :attribute may not end with one of the following: :values.',
+    'doesnt_start_with' => 'The :attribute may not start with one of the following: :values.',
+    'email'             => 'The :attribute must be a valid email address.',
+    'ends_with'         => 'The :attribute must end with one of the following: :values.',
+    'enum'              => 'The selected :attribute is invalid.',
+    'exists'            => 'The selected :attribute is invalid.',
+    'file'              => 'The :attribute must be a file.',
+    'filled'            => 'The :attribute field must have a value.',
+    'gt'                => [
         'array'   => 'The :attribute must have more than :value items.',
         'file'    => 'The :attribute must be greater than :value kilobytes.',
         'numeric' => 'The :attribute must be greater than :value.',
         'string'  => 'The :attribute must be greater than :value characters.',
     ],
-    'gte'                  => [
+    'gte' => [
         'array'   => 'The :attribute must have :value items or more.',
         'file'    => 'The :attribute must be greater than or equal to :value kilobytes.',
         'numeric' => 'The :attribute must be greater than or equal to :value.',
         'string'  => 'The :attribute must be greater than or equal to :value characters.',
     ],
-    'image'                => 'The :attribute must be an image.',
-    'in'                   => 'The selected :attribute is invalid.',
-    'in_array'             => 'The :attribute field does not exist in :other.',
-    'integer'              => 'The :attribute must be an integer.',
-    'ip'                   => 'The :attribute must be a valid IP address.',
-    'ipv4'                 => 'The :attribute must be a valid IPv4 address.',
-    'ipv6'                 => 'The :attribute must be a valid IPv6 address.',
-    'json'                 => 'The :attribute must be a valid JSON string.',
-    'lt'                   => [
+    'image'    => 'The :attribute must be an image.',
+    'in'       => 'The selected :attribute is invalid.',
+    'in_array' => 'The :attribute field does not exist in :other.',
+    'integer'  => 'The :attribute must be an integer.',
+    'ip'       => 'The :attribute must be a valid IP address.',
+    'ipv4'     => 'The :attribute must be a valid IPv4 address.',
+    'ipv6'     => 'The :attribute must be a valid IPv6 address.',
+    'json'     => 'The :attribute must be a valid JSON string.',
+    'lt'       => [
         'array'   => 'The :attribute must have less than :value items.',
         'file'    => 'The :attribute must be less than :value kilobytes.',
         'numeric' => 'The :attribute must be less than :value.',
         'string'  => 'The :attribute must be less than :value characters.',
     ],
-    'lte'                  => [
+    'lte' => [
         'array'   => 'The :attribute must not have more than :value items.',
         'file'    => 'The :attribute must be less than or equal to :value kilobytes.',
         'numeric' => 'The :attribute must be less than or equal to :value.',
         'string'  => 'The :attribute must be less than or equal to :value characters.',
     ],
-    'mac_address'          => 'The :attribute must be a valid MAC address.',
-    'max'                  => [
+    'mac_address' => 'The :attribute must be a valid MAC address.',
+    'max'         => [
         'array'   => 'The :attribute must not have more than :max items.',
         'file'    => 'The :attribute must not be greater than :max kilobytes.',
         'numeric' => 'The :attribute must not be greater than :max.',
         'string'  => 'The :attribute must not be greater than :max characters.',
     ],
-    'max_digits'           => 'The :attribute must not have more than :max digits.',
-    'mimes'                => 'The :attribute must be a file of type: :values.',
-    'mimetypes'            => 'The :attribute must be a file of type: :values.',
-    'min'                  => [
+    'max_digits' => 'The :attribute must not have more than :max digits.',
+    'mimes'      => 'The :attribute must be a file of type: :values.',
+    'mimetypes'  => 'The :attribute must be a file of type: :values.',
+    'min'        => [
         'array'   => 'The :attribute must have at least :min items.',
         'file'    => 'The :attribute must be at least :min kilobytes.',
         'numeric' => 'The :attribute must be at least :min.',
         'string'  => 'The :attribute must be at least :min characters.',
     ],
-    'min_digits'           => 'The :attribute must have at least :min digits.',
-    'multiple_of'          => 'The :attribute must be a multiple of :value.',
-    'not_in'               => 'The selected :attribute is invalid.',
-    'not_regex'            => 'The :attribute format is invalid.',
-    'numeric'              => 'The :attribute must be a number.',
-    'password'             => [
+    'min_digits'  => 'The :attribute must have at least :min digits.',
+    'multiple_of' => 'The :attribute must be a multiple of :value.',
+    'not_in'      => 'The selected :attribute is invalid.',
+    'not_regex'   => 'The :attribute format is invalid.',
+    'numeric'     => 'The :attribute must be a number.',
+    'password'    => [
         'letters'       => 'The :attribute must contain at least one letter.',
         'mixed'         => 'The :attribute must contain at least one uppercase and one lowercase letter.',
         'numbers'       => 'The :attribute must contain at least one number.',
@@ -131,13 +131,13 @@ return [
         'numeric' => 'The :attribute must be :size.',
         'string'  => 'The :attribute must be :size characters.',
     ],
-    'starts_with'          => 'The :attribute must start with one of the following: :values.',
-    'string'               => 'The :attribute must be a string.',
-    'timezone'             => 'The :attribute must be a valid timezone.',
-    'unique'               => 'The :attribute has already been taken.',
-    'uploaded'             => 'The :attribute failed to upload.',
-    'url'                  => 'The :attribute must be a valid URL.',
-    'uuid'                 => 'The :attribute must be a valid UUID.',
+    'starts_with' => 'The :attribute must start with one of the following: :values.',
+    'string'      => 'The :attribute must be a string.',
+    'timezone'    => 'The :attribute must be a valid timezone.',
+    'unique'      => 'The :attribute has already been taken.',
+    'uploaded'    => 'The :attribute failed to upload.',
+    'url'         => 'The :attribute must be a valid URL.',
+    'uuid'        => 'The :attribute must be a valid UUID.',
 
     /*
     |--------------------------------------------------------------------------
@@ -150,7 +150,7 @@ return [
     |
     */
 
-    'custom'               => [
+    'custom' => [
         'attribute-name' => [
             'rule-name' => 'custom-message',
         ],
@@ -167,5 +167,5 @@ return [
     |
     */
 
-    'attributes'           => [],
+    'attributes' => [],
 ];
