@@ -1,9 +1,21 @@
 <?php
 
 return [
+    /*
+    |--------------------------------------------------------------------------
+    | Password Reset Language Lines
+    |--------------------------------------------------------------------------
+    |
+    | The following language lines are the default lines which match reasons
+    | that are given by the password broker for a password update attempt
+    | has failed, such as for an invalid token or invalid new password.
+    |
+    */
+
+    'password'  => 'A senha e a confirmação devem combinar e possuir pelo menos seis caracteres.',
     'reset'     => 'Sua senha foi redefinida!',
-    'sent'      => 'Enviamos um link para redefinir a sua senha por e-mail.',
-    'throttled' => 'Por favor espere antes de tentar novamente.',
-    'token'     => 'Esse código de redefinição de senha é inválido.',
-    'user'      => 'Não conseguimos encontrar nenhum usuário com o endereço de e-mail informado.',
+    'sent'      => 'Enviamos seu link de redefinição de senha por e-mail!',
+    'throttled' => 'Aguarde antes de tentar novamente.',
+    'token'     => 'Este token de redefinição de senha é inválido.',
+    'user'      => 'Não encontramos um usuário com esse endereço de e-mail.',
 ];
